compiletest: localize `compile_test_and_save_assembly` to assembly test module
